Move common schema types to a separate gradle submodule
