+++ ballerina/Ballerina.toml
@@ -19,4 +19,4 @@ path = "../native/build/libs/graphql-native-1.5.0-SNAPSHOT.jar"
 groupId = "io.ballerina.stdlib"
 artifactId = "graphql-compiler-plugin"
 version = "1.5.0"
-path = "../compiler-plugin/build/libs/graphql-compiler-plugin-1.5.0-SNAPSHOT.jar"
+path = "../commons/build/libs/graphql-commons-1.5.0-SNAPSHOT.jar"
