+++ ballerina/CompilerPlugin.toml
@@ -4,3 +4,6 @@ class = "io.ballerina.stdlib.graphql.compiler.GraphqlCompilerPlugin"
 
 [[dependency]]
 path = "../compiler-plugin/build/libs/graphql-compiler-plugin-1.5.0-SNAPSHOT.jar"
+
+[[dependency]]
+path = "../commons/build/libs/graphql-commons-1.5.0-SNAPSHOT.jar"
