+++ ballerina/build.gradle
@@ -112,8 +112,10 @@ updateTomlFiles.dependsOn copyStdlibs
 build.dependsOn "generatePomFileForMavenPublication"
 build.dependsOn ":${packageName}-native:build"
 build.dependsOn ":${packageName}-compiler-plugin:build"
+build.dependsOn ":${packageName}-commons:build"
 test.dependsOn ":${packageName}-native:build"
 test.dependsOn ":${packageName}-compiler-plugin:build"
+test.dependsOn ":${packageName}-commons:build"
 
 publishToMavenLocal.dependsOn build
 publish.dependsOn build
