+++ build-config/resources/Ballerina.toml
@@ -19,4 +19,4 @@ path = "../native/build/libs/graphql-native-@project.version@.jar"
 groupId = "io.ballerina.stdlib"
 artifactId = "graphql-compiler-plugin"
 version = "@toml.version@"
-path = "../compiler-plugin/build/libs/graphql-compiler-plugin-@project.version@.jar"
+path = "../commons/build/libs/graphql-commons-@project.version@.jar"
