+++ build-config/resources/CompilerPlugin.toml
@@ -4,3 +4,6 @@ class = "io.ballerina.stdlib.graphql.compiler.GraphqlCompilerPlugin"
 
 [[dependency]]
 path = "../compiler-plugin/build/libs/graphql-compiler-plugin-@project.version@.jar"
+
+[[dependency]]
+path = "../commons/build/libs/graphql-commons-@project.version@.jar"
