+++ commons/src/main/java/io/ballerina/stdlib/graphql/commons/Utils.java
@@ -1,0 +1,78 @@
+/*
+ * Copyright (c) 2022, WSO2 LLC. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerina.stdlib.graphql.commons;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+/**
+ * Utility class for Ballerina GraphQL schema types.
+ */
+public class Utils {
+    private static final String UNICODE_REGEX = "\\\\(\\\\*)u\\{([a-fA-F0-9]+)\\}";
+    private static final Pattern UNICODE_PATTERN = Pattern.compile(UNICODE_REGEX);
+
+    private static final String SINGLE_QUOTE_CHARACTER = "'";
+
+    public static String removeEscapeCharacter(String identifier) {
+        if (identifier == null) {
+            return null;
+        }
+
+        Matcher matcher = UNICODE_PATTERN.matcher(identifier);
+        StringBuffer buffer = new StringBuffer(identifier.length());
+        while (matcher.find()) {
+            String leadingSlashes = matcher.group(1);
+            if (isEscapedNumericEscape(leadingSlashes)) {
+                // e.g. \\u{61}, \\\\u{61}
+                continue;
+            }
+
+            int codePoint = Integer.parseInt(matcher.group(2), 16);
+            char[] chars = Character.toChars(codePoint);
+            String ch = String.valueOf(chars);
+
+            if (ch.equals("\\")) {
+                // Ballerina string unescaping is done in two stages.
+                // 1. unicode code point unescaping (doing separately as [2] does not support code points > 0xFFFF)
+                // 2. java unescaping
+                // Replacing unicode code point of backslash at [1] would compromise [2]. Therefore, special case it.
+                matcher.appendReplacement(buffer, Matcher.quoteReplacement(leadingSlashes + "\\u005C"));
+            } else {
+                matcher.appendReplacement(buffer, Matcher.quoteReplacement(leadingSlashes + ch));
+            }
+        }
+        matcher.appendTail(buffer);
+        String value = String.valueOf(buffer);
+
+        if (value.startsWith(SINGLE_QUOTE_CHARACTER)) {
+            return value.substring(1);
+        }
+        return value;
+    }
+
+    private static boolean isEscapedNumericEscape(String leadingSlashes) {
+        return !isEven(leadingSlashes.length());
+    }
+
+    private static boolean isEven(int n) {
+        // (n & 1) is 0 when n is even.
+        return (n & 1) == 0;
+    }
+}
