+++ commons/src/main/java/io/ballerina/stdlib/graphql/commons/types/Description.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.graphql.compiler.schema.types;
+package io.ballerina.stdlib.graphql.commons.types;
 
 /**
  * Stores descriptions for the default types in a GraphQL schema.
