+++ commons/src/main/java/io/ballerina/stdlib/graphql/commons/types/Type.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.graphql.compiler.schema.types;
+package io.ballerina.stdlib.graphql.commons.types;
 
 import java.io.Serializable;
 import java.util.ArrayList;
@@ -25,7 +25,7 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
-import static io.ballerina.stdlib.graphql.compiler.schema.generator.GeneratorUtils.removeEscapeCharacter;
+import static io.ballerina.stdlib.graphql.commons.Utils.removeEscapeCharacter;
 
 /**
  * Represents the {@code __Type} type in GraphQL schema.
