+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/schema/generator/GeneratorUtils.java
@@ -20,14 +20,12 @@ package io.ballerina.stdlib.graphql.compiler.schema.generator;
 
 import io.ballerina.compiler.api.symbols.Documentable;
 import io.ballerina.compiler.api.symbols.TypeSymbol;
-import io.ballerina.stdlib.graphql.compiler.schema.types.ScalarType;
-import io.ballerina.stdlib.graphql.compiler.schema.types.Type;
-import io.ballerina.stdlib.graphql.compiler.schema.types.TypeKind;
+import io.ballerina.stdlib.graphql.commons.types.ScalarType;
+import io.ballerina.stdlib.graphql.commons.types.Type;
+import io.ballerina.stdlib.graphql.commons.types.TypeKind;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * Utility methods for Ballerina GraphQL schema generator.
@@ -35,11 +33,6 @@ import java.util.regex.Pattern;
 public class GeneratorUtils {
     private GeneratorUtils() {
     }
-
-    private static final String UNICODE_REGEX = "\\\\(\\\\*)u\\{([a-fA-F0-9]+)\\}";
-    private static final Pattern UNICODE_PATTERN = Pattern.compile(UNICODE_REGEX);
-
-    private static final String SINGLE_QUOTE_CHARACTER = "'";
 
     public static final String UNION_TYPE_NAME_DELIMITER = "_";
 
@@ -98,52 +91,6 @@ public class GeneratorUtils {
             return null;
         }
         return documentable.documentation().get().deprecatedDescription().get();
-    }
-
-    public static String removeEscapeCharacter(String identifier) {
-        if (identifier == null) {
-            return null;
-        }
-
-        Matcher matcher = UNICODE_PATTERN.matcher(identifier);
-        StringBuffer buffer = new StringBuffer(identifier.length());
-        while (matcher.find()) {
-            String leadingSlashes = matcher.group(1);
-            if (isEscapedNumericEscape(leadingSlashes)) {
-                // e.g. \\u{61}, \\\\u{61}
-                continue;
-            }
-
-            int codePoint = Integer.parseInt(matcher.group(2), 16);
-            char[] chars = Character.toChars(codePoint);
-            String ch = String.valueOf(chars);
-
-            if (ch.equals("\\")) {
-                // Ballerina string unescaping is done in two stages.
-                // 1. unicode code point unescaping (doing separately as [2] does not support code points > 0xFFFF)
-                // 2. java unescaping
-                // Replacing unicode code point of backslash at [1] would compromise [2]. Therefore, special case it.
-                matcher.appendReplacement(buffer, Matcher.quoteReplacement(leadingSlashes + "\\u005C"));
-            } else {
-                matcher.appendReplacement(buffer, Matcher.quoteReplacement(leadingSlashes + ch));
-            }
-        }
-        matcher.appendTail(buffer);
-        String value = String.valueOf(buffer);
-
-        if (value.startsWith(SINGLE_QUOTE_CHARACTER)) {
-            return value.substring(1);
-        }
-        return value;
-    }
-
-    private static boolean isEscapedNumericEscape(String leadingSlashes) {
-        return !isEven(leadingSlashes.length());
-    }
-
-    private static boolean isEven(int n) {
-        // (n & 1) is 0 when n is even.
-        return (n & 1) == 0;
     }
 
     public static Type getWrapperType(Type type, TypeKind typeKind) {
