+++ examples/build.gradle
@@ -1,5 +1,3 @@
-import org.apache.tools.ant.taskdefs.condition.Os
-
 /*
  * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
@@ -17,6 +15,8 @@ import org.apache.tools.ant.taskdefs.condition.Os
  * specific language governing permissions and limitations
  * under the License.
  */
+
+import org.apache.tools.ant.taskdefs.condition.Os
 
 apply plugin: 'java'
