+++ native/build.gradle
@@ -29,7 +29,7 @@ dependencies {
 
     implementation group: 'org.ballerinalang', name: 'ballerina-lang', version: "${ballerinaLangVersion}"
     implementation group: 'org.ballerinalang', name: 'ballerina-runtime', version: "${ballerinaLangVersion}"
-    implementation project(':graphql-compiler-plugin')
+    implementation project(':graphql-commons')
 }
 
 def excludePattern = '**/module-info.java'
