+++ native/src/main/java/module-info.java
@@ -18,5 +18,5 @@
 
 module io.ballerina.stdlib.graphql.runtime {
     requires io.ballerina.runtime;
-    requires io.ballerina.stdlib.graphql.compiler;
+    requires io.ballerina.stdlib.graphql.commons;
 }
