+++ settings.gradle
@@ -29,6 +29,7 @@ include ':graphql-ballerina'
 include ':graphql-compiler-plugin-tests'
 include ':graphql-ballerina-tests'
 include ':graphql-examples'
+include ':graphql-commons'
 
 project(':checkstyle').projectDir = file("build-config${File.separator}checkstyle")
 project(':graphql-native').projectDir = file("native")
@@ -37,6 +38,7 @@ project(':graphql-ballerina').projectDir = file("ballerina")
 project(':graphql-compiler-plugin-tests').projectDir = file("compiler-plugin-tests")
 project(':graphql-ballerina-tests').projectDir = file("ballerina-tests")
 project(':graphql-examples').projectDir = file("examples")
+project(':graphql-commons').projectDir = file("commons")
 
 gradleEnterprise {
    buildScan {
